refine inject_mode
